20210814 Minor Changes2
More minor changes. This should be everything for the players to properly choose their teams in real time. Might need to implement some sort of time delay to keep players from spamming buttons and causing issues, but for the time being, this is good enough. Next it starting a game.